Refactor: Removed theme
Removed theme feature, didn't bode well with the theme i have in mind
