+++ src/views/PhotoModal.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="photo-modal" v-if="photo">
-    <svg class="w-6 h-6 close" @click="redirect" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z"></path></svg>
+    <svg class="w-6 h-6 close" @click="redirect" fill="#fff" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z"></path></svg>
     <div class="photo-modal-content">
       <img :src="photo.urls.small" alt="">
       <div class="photo-body">
@@ -86,8 +86,9 @@ export default {
     width: 450px;
   }
   .close {
-    /* top: 50px;
-    right: 70px; */
+    top: 30px;
+    right: 70px;
+    width: 30px;
   }
 }
 </style>
